Move checkpoint_utils to common location for other model using it
